Split out the tale tab selector into a set of smaller components
