+++ app/components/ui/tale-tab-interact/component.js
@@ -1,0 +1,12 @@
+import Ember from 'ember';
+
+
+export default Ember.Component.extend({
+  model: null,
+  init() {
+    this._super(...arguments);
+  },
+  
+  actions: {
+  }
+});
