+++ app/components/ui/tale-tab-interact/template.hbs
@@ -1,0 +1,21 @@
+{{#if (not-eq model.status 0)}}
+    {{#unless model.iframe}}
+        <div class="placeholder message">
+            <p>The compute environment for this Tale must run in a separate browser window.<br/>
+              <a class="ui primary button huge" href="{{model.url}}" target="_blank">Go to new window</a></p>
+
+        </div>
+    {{else}}
+      <div class="rounded-bottom" style="padding: 0; margin: 0; overflow-y: hidden;">
+        <iframe id="frontendDisplay"
+                src="{{model.url}}"
+                style="padding: 0; width:100%; height: inherit;"
+                class="ui basic segment">
+        </iframe>
+      </div>
+    {{/unless}}
+{{else}}
+  <div class="placeholder message">
+    <p>Creating your new Tale...</p>
+  </div>
+{{/if}}
