+++ app/components/ui/tale-tabs-selector/component.js
@@ -1,0 +1,30 @@
+import Ember from 'ember';
+
+
+export default Ember.Component.extend({
+  init() {
+    this._super(...arguments);
+    
+    this.set("activeTabInteract", true);
+    this.set("activeTabFiles", false);
+    this.set("activeTabMetadata", false);
+  },
+  
+  actions: {
+    activateInteract() {
+      this.set("activeTabInteract", true);
+      this.set("activeTabFiles", false);
+      this.set("activeTabMetadata", false);
+    },
+    activateFiles() {
+      this.set("activeTabInteract", false);
+      this.set("activeTabFiles", true);
+      this.set("activeTabMetadata", false);
+    },
+    activateMetadata() {
+      this.set("activeTabInteract", false);
+      this.set("activeTabFiles", false);
+      this.set("activeTabMetadata", true);
+    },
+  }
+});
